+++ 2022/02/main.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"log"
 
-	aoc "github.com/matthewchivers/advent-of-code/utils"
+	aoc "github.com/matthewchivers/advent-of-code/util"
 )
 
 var (
