+++ 2022/06/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"log"
 
-	aoc "github.com/matthewchivers/advent-of-code/utils"
+	aoc "github.com/matthewchivers/advent-of-code/util"
 )
 
 type lastN struct {
